stripped out reductions, subspaces and eigen
